requirements.txt
python-telegram-bot>=20,<22
requests
pandas
ta
python-dotenv
